Initial Commit
Disini semua bagian masih bekerja dengan sangat baik , hanya saja masih belum bisa menggunakan routing app.post('/predict')

+++ src/index.js
@@ -8,6 +8,11 @@ const saltRounds = 10;
 const { createAccessToken, createRefreshToken, sendAccessToken, sendRefreshToken } = require('./tokens.js');
 const { isAuth } = require('./isAuth.js');
 const port = 3000;
+const Sequelize = require('sequelize');
+const Op = Sequelize.Op;
+const util = require('util');
+const Multer = require('multer');
+const { Storage } = require('@google-cloud/storage');
 
 const app = express();
 
@@ -17,12 +22,19 @@ app.use(express.urlencoded({ extended: false }));
 
 const User = require('../model/user.js');
 const Blog = require('../model/blog.js');
+const History = require('../model/history.js');
+const { error } = require('console');
 
 // One to many relationship
 User.hasMany(Blog, {
   foreignKey: 'userId',
 });
 Blog.belongsTo(User);
+
+User.hasMany(History, {
+  foreignKey: 'userId',
+});
+History.belongsTo(User);
 
 // Synchronize the database with the model
 db.sync({})
@@ -32,6 +44,19 @@ db.sync({})
   .catch((err) => {
     console.log(err);
   });
+
+const dateMonthYear = () => {
+  const today = new Date();
+  const yyyy = today.getFullYear();
+  let mm = today.getMonth() + 1; // Months start at 0!
+  let dd = today.getDate();
+
+  if (dd < 10) dd = '0' + dd;
+  if (mm < 10) mm = '0' + mm;
+
+  const formattedToday = dd + '/' + mm + '/' + yyyy;
+  return formattedToday;
+};
 
 // Register a user and password
 app.post('/register', async (req, res) => {
@@ -54,6 +79,7 @@ app.post('/register', async (req, res) => {
         res.status(200);
         res.json({
           error: false,
+          message: 'User berhasil dibuat',
           response: result,
         });
       });
@@ -62,7 +88,7 @@ app.post('/register', async (req, res) => {
     res.status(403);
     res.json({
       error: true,
-      response: `${err.message}`,
+      message: `${err.message}`,
     });
   }
 });
@@ -87,13 +113,22 @@ app.post('/login', async (req, res) => {
     await getData.save();
 
     sendRefreshToken(res, refreshToken);
-    sendAccessToken(req, res, accessToken);
 
     res.status(200);
+    res.json({
+      error: false,
+      message: 'Login Sukses',
+      loginResult: {
+        userId: getData.id,
+        name: getData.fullName,
+        token: accessToken,
+      },
+    });
   } catch (err) {
     res.status(403);
     res.json({
-      error: `${err.message}`,
+      error: true,
+      message: `${err.message}`,
     });
   }
 });
@@ -103,89 +138,189 @@ app.post('/logout', (req, res) => {
   res.clearCookie('refreshToken', { path: '/refresh_token' });
   res.status(200);
   return res.send({
-    message: 'Logged Out',
+    message: 'Berhasil Keluar',
   });
-});
-
-// Accessing protected route , with Accesstoken , just a prototype
-app.post('/protected', isAuth, async (req, res) => {
-  try {
-    const id = req.user.id;
-    if (id !== null) {
-      res.send({
-        data: 'This is protected data',
-      });
-      res.status(200);
-    }
-  } catch (err) {
-    res.status(403);
-    res.send({
-      error: `${err.message}`,
-    });
-  }
 });
 
 // Getting all blogs with accessToken
 app.get('/blogs', isAuth, async (req, res) => {
-  try {
-    const blogs = await Blog.findAll();
-    const arrayTemp = [];
-    blogs.forEach((blog) => {
-      arrayTemp.push(blog);
-    });
-    res.status(200);
-    res.send(arrayTemp);
-  } catch (err) {
-    res.status(403);
-    res.send({
-      err: `${err.message}`,
-    });
-  }
+  const blogs = await Blog.findAll();
+  const arrayTemp = [];
+  blogs.forEach((blog) => {
+    arrayTemp.push(blog);
+  });
+  res.status(200);
+  res.send({
+    error: false,
+    message: 'Blogs berhasil dipanggil',
+    blogs: arrayTemp,
+  });
 });
 
 // Getting all blogs that filtered with the user's id
 app.get('/userPrivateBlogs', isAuth, async (req, res) => {
+  const id = req.user.id;
+  const arrayTemp = [];
+  const userPrivateBlogs = await Blog.findAll({
+    where: {
+      userId: id,
+    },
+  });
+  userPrivateBlogs.forEach((blogFoundByUserId) => {
+    arrayTemp.push(blogFoundByUserId);
+  });
+  res.status(200);
+  res.send({
+    error: false,
+    message: 'Blogs berhasil dipanggil',
+    blogs: arrayTemp,
+  });
+});
+
+// Getting blog by using blogId
+app.get('/blogs/:blogId', isAuth, async (req, res) => {
   try {
-    const id = req.user.id;
-    const arrayTemp = [];
-    const userPrivateBlogs = await Blog.findAll({
-      where: {
-        userId: id,
-      },
+    const selectedBlog = await Blog.findOne({
+      where: { blogId: req.params.blogId },
     });
-    userPrivateBlogs.forEach((blogFoundByUserId) => {
-      arrayTemp.push(blogFoundByUserId);
-    });
+    if (!selectedBlog) {
+      throw new Error('Blog yang mau dibuka tidak ada');
+    }
     res.status(200);
-    res.send(arrayTemp);
+    res.send({
+      error: false,
+      message: 'Blog berhasil dibuka',
+      response: selectedBlog,
+    });
   } catch (err) {
     res.status(403);
     res.send({
-      err: `${err.message}`,
+      error: true,
+      message: `${err.message}`,
     });
   }
 });
 
+// Getting blog while searching for name
+app.get('/blogsWithName', isAuth, async (req, res) => {
+  try {
+    const blogTitle = req.query.blogTitle;
+
+    const selectedBlogs = await Blog.findAll({
+      where: {
+        blogTitle: {
+          [Op.like]: `%${blogTitle}%`,
+        },
+      },
+    });
+
+    if (!selectedBlogs) {
+      throw new Error('Blog yang dicari tidak ada');
+    }
+    res.status(200);
+    res.send({
+      error: false,
+      message: 'Blog berhasil dibuka',
+      response: selectedBlogs,
+    });
+  } catch (err) {
+    res.status(403);
+    res.send({
+      error: true,
+      message: `${err.message}`,
+    });
+  }
+});
+
+// Connect to the cloud storage with the service account
+const storage = new Storage({
+  projectId: 'something-idk-388806',
+  keyFilename: 'service_account.json',
+});
+
+const bucketName = 'capstone-blog-bucket';
+const bucket = storage.bucket(bucketName);
+
+let processFile = Multer({
+  storage: Multer.memoryStorage(),
+}).single('file');
+
+let processFileMiddleware = util.promisify(processFile);
+
 // Creating a blog with accessToken
 app.post('/blogs', isAuth, async (req, res) => {
   try {
-    const blogTitle = req.body.blogTitle;
-    const blogDescription = req.body.blogDescription;
-    const userId = req.user.id;
-    if (userId !== null) {
-      await Blog.create({
-        blogTitle: blogTitle,
-        blogDescription: blogDescription,
-        userId: userId,
-      });
-      res.status(200);
-      res.send({
-        msg: 'Blog created successfully',
-      });
-    } else {
-      res.status(403);
-      return new Error('You need to login');
+    await processFileMiddleware(req, res);
+
+    if (!req.file) {
+      return res.status(400).send({ message: 'Please upload a file!' });
     }
+
+    const blob = bucket.file(req.file.originalname);
+    const blobStream = blob.createWriteStream();
+
+    blobStream.on('NotFoundError', (err) => {
+      res.status(500).send({ message: err.message });
+    });
+
+    blobStream.on('InvalidRequestError', (err) => {
+      res.status(500).send({ message: err.message });
+    });
+
+    blobStream.on('finish', async () => {
+      const publicUrl = `https://storage.googleapis.com/${bucketName}/${blob.name}`;
+      try {
+        await bucket.file(req.file.originalname).makePublic();
+      } catch {
+        return res.status(500).send({
+          message: `Uploaded the file successfully: ${req.file.originalname}, but public access is denied`,
+          url: publicUrl,
+        });
+      }
+
+      const blogTitle = req.body.blogTitle;
+      const blogDescription = req.body.blogDescription;
+      const userId = req.user.id;
+      const userFound = await User.findOne({
+        where: { id: userId },
+      });
+      const createdBy = userFound.fullName;
+
+      if (!blogTitle) {
+        return res.status(500).send({ message: 'Masukkan judul blog' });
+      }
+      if (!blogDescription) {
+        return res.status(500).send({ message: 'Masukkan isi dari blog' });
+      }
+
+      // Pakai kode ini error
+      // if (!blogTitle) {
+      //   throw new Error('Masukkan judul blog');
+      // }
+
+      // if (!blogDescription) {
+      //   throw new Error('Masukkan isi dari blog');
+      // }
+
+      if (userId !== null) {
+        await Blog.create({
+          blogTitle: blogTitle,
+          blogDescription: blogDescription,
+          dateCreated: dateMonthYear(),
+          createdBy: createdBy,
+          imageUrl: publicUrl,
+          userId: userId,
+        });
+        res.status(200);
+        res.send({
+          message: 'Blog berhasil dibuat',
+        });
+      } else {
+        res.status(403);
+        throw new Error('Anda perlu login');
+      }
+    });
+    blobStream.end(req.file.buffer);
   } catch (err) {
     res.status(403);
     res.send({
@@ -195,25 +330,33 @@ app.post('/blogs', isAuth, async (req, res) => {
 });
 
 // Deleting a blog that user created using path parameter
-app.delete('/blogs/:blogId', (req, res) => {
+app.delete('/blogs/:blogId', isAuth, async (req, res) => {
   try {
-    Blog.destroy({
+    const selectedBlog = await Blog.findOne({
+      where: { blogId: req.params.blogId },
+    });
+    if (!selectedBlog) {
+      throw new Error('Blog yang mau dihapus tidak ada');
+    }
+    await Blog.destroy({
       where: { blogId: req.params.blogId },
     });
     res.status(200);
     res.send({
-      msg: 'Blog deleted successfully',
+      error: false,
+      message: 'Blog berhasil dihapus',
     });
   } catch (err) {
     res.status(403);
     res.send({
-      err: `${err.message}`,
+      error: true,
+      message: `${err.message}`,
     });
   }
 });
 
 // Asking for refreshToken (but not necesary here)
-app.post('/refresh_token', async (req, res) => {
+app.post('/refresh_token', isAuth, async (req, res) => {
   const token = req.cookies.refreshtoken;
 
   if (!token) return res.send({ accessToken: '' });
@@ -239,6 +382,22 @@ app.post('/refresh_token', async (req, res) => {
   sendRefreshToken(res, refreshToken);
   await user.save();
   return res.send({ accessToken: accessToken });
+});
+
+// Adding the picture to the ML Model
+app.post('/predict', async (req, res) => {
+  try {
+    let formData = new FormData(req.file);
+    formData.append('file', req.file);
+    const response = await fetch('https://hydroplant-glnz5e5urq-et.a.run.app/predict', {
+      method: 'POST',
+      body: formData,
+    });
+    const result = await response.json();
+    console.log('Success: ', result);
+  } catch (err) {
+    console.error('Error: ', err);
+  }
 });
 
 app.listen(port, () => {
